+++ TeamLegend/TeamLegend.Services/TeamsService.cs
@@ -21,6 +21,9 @@
 
         public async Task<Team> CreateAsync(Team team)
         {
+            if (team == null)
+                return null;
+
             await this.context.Teams.AddAsync(team);
             await this.context.SaveChangesAsync();
 
@@ -30,9 +33,7 @@
         public async Task<Team> GetByIdAsync(string id)
         {
             if (string.IsNullOrEmpty(id))
-            {
                 return null;
-            }
 
             return await this.context.Teams.FirstOrDefaultAsync(p => p.Id == id);
         }
@@ -40,9 +41,7 @@
         public async Task<Team> GetByNameAsync(string name)
         {
             if (string.IsNullOrEmpty(name))
-            {
                 return null;
-            }
 
             return await this.context.Teams.FirstOrDefaultAsync(p => p.Name == name);
         }
@@ -57,15 +56,11 @@
             return await this.context.Teams.Where(t => t.LeagueId == null).ToListAsync();
         }
 
-        public async Task<Team> UpdateAsync(Team team)
-        {
-            this.context.Teams.Update(team);
-            await this.context.SaveChangesAsync();
-            return team;
-        }
-
         public async Task<bool> DeleteAsync(Team team)
         {
+            if (team == null)
+                return false;
+
             this.context.Teams.Remove(team);
             await this.context.SaveChangesAsync();
 
@@ -74,6 +69,9 @@
 
         public async Task<Team> SetStadiumAsync(Team team, Stadium stadium)
         {
+            if (team == null || stadium == null)
+                return null;
+
             team.Stadium = stadium;
             this.context.Teams.Update(team);
             await this.context.SaveChangesAsync();
@@ -83,6 +81,9 @@
 
         public async Task<Team> AddNewPlayersAsync(Team team, List<Player> playersToAdd)
         {
+            if (team == null || playersToAdd == null)
+                return null;
+
             playersToAdd.ForEach(p => team.Players.Add(p));
             this.context.Teams.Update(team);
 
@@ -93,6 +94,9 @@
 
         public async Task<Team> AddManagerAsync(Team team, Manager manager)
         {
+            if (team == null || manager == null)
+                return null;
+
             team.Manager = manager;
             this.context.Teams.Update(team);
